New Update

+++ core-service/src/main/resources/db/migration/V1__init.sql
@@ -12,36 +12,36 @@ create table products
     id              bigserial primary key,
     title           varchar(255),
     category_id     bigint references categories (id),
-    price           int,
+    price           numeric(8, 2),
     created_at      timestamp default current_timestamp,
     updated_at      timestamp default current_timestamp
 );
 
 insert into products (title, price, category_id) values
-('Milk', 80, 1),
-('Milk1', 22, 1),
-('Milk2', 83, 1),
-('Milk3', 85, 1),
-('Milk4', 70, 1),
-('Milk5', 75, 1),
-('Milk6', 86, 1),
-('Milk7', 44, 1),
-('Milk8', 56, 1),
-('Milk9', 83, 1),
-('Milk11', 55, 1),
-('Milk22', 74, 1),
-('Milk33', 36, 1),
-('Milk44', 74, 1),
-('Milk55', 86, 1),
-('Milk77', 11, 1),
-('Bread', 25, 1),
-('Cheese', 300, 1);
+('Milk', 80.00, 1),
+('Milk1', 22.00, 1),
+('Milk2', 83.00, 1),
+('Milk3', 85.00, 1),
+('Milk4', 70.00, 1),
+('Milk5', 75.00, 1),
+('Milk6', 86.00, 1),
+('Milk7', 44.00, 1),
+('Milk8', 56.00, 1),
+('Milk9', 83.00, 1),
+('Milk11', 55.00, 1),
+('Milk22', 74.00, 1),
+('Milk33', 36.00, 1),
+('Milk44', 74.00, 1),
+('Milk55', 86.00, 1),
+('Milk77', 11.00, 1),
+('Bread', 25.00, 1),
+('Cheese', 300.00, 1);
 
 create table orders
 (
     id          bigserial primary key,
     username    varchar(255) not null,
-    total_price int not null,
+    total_price numeric(8, 2) not null,
     address     varchar(255),
     phone       varchar(255),
     created_at  timestamp default current_timestamp,
@@ -54,8 +54,8 @@ create table order_items
     product_id        bigint not null references products (id),
     order_id          bigint not null references orders (id),
     quantity          int    not null,
-    price_per_product int    not null,
-    price             int    not null,
+    price_per_product numeric(8, 2)    not null,
+    price             numeric(8, 2)    not null,
     created_at        timestamp default current_timestamp,
     updated_at        timestamp default current_timestamp
 );
